Fix T-lang triage agenda by trimming down FCP template
For some as-yet undiagnosed reason, the T-lang triage agenda generator
has stopped working.  It fails when populating the template due to a
missing variable.  Presumably something has changed on the GitHub API
or in how issues are handled that is causing this variable to not be
populated.

For the moment, to keep things working, let's just trim down the
template to exclude details about each FCP.