Add (`T-lang`, `minutes`) tags to T-lang agendas
It's useful to have all documents for T-lang include a `T-lang` tag so
that these documents can be identified when moved or copied to places
that include content from other teams.  Let's include `T-lang` in the
set of tags for the T-lang agendas.

Similarly, it's useful to have all documents that contain the minutes
of a meeting tagged with `minutes`, so let's do that here as well.