Consolidate nomination sections in T-lang triage
Previously we had two sections for nominated items.  We had a section
for those discussed during the meeting, and a section for those not
discussed during the meeting.  The idea was that items were selected
during the meeting and moved from one section to another.

Lately, however, we've been organizing what should be discussed ahead
of the meeting and putting the items in a likely order.  We then tend
to go through them in that order.  Not moving things around during the
meeting helps things to go a bit faster.

Done this way, it makes less sense to have two sections.  We simply go
through the nominated items until we run out of time, and then we note
in the minutes where the meeting ends.  So let's pare the nominated
items down to one section rather than two.