Render FCP links consistently with other links
We were rendering the links to issues for which there is a proposed
FCP as an element in an unordered list, but we were rendering other
links to issues simply as a paragraph element.  Let's make this
consistent by rendering the links to proposed FCP issues in the same
way as the links to other issues.