Refactor `project_number` handling for design meetings
When we fetch project items by status (which we currently only do to
assemble the list of design meetings), we need to include the
`project_number` as one of the arguments in the query.  This
`project_number` corresponds to a "project" in GitHub, which is a kind
of spreadsheet.  Each design meeting schedule would be a different
"project" in this sense and would have a different `project_number`.

This `project_number` was hardcoded in the function that queries for
the project items by status.

So that we can use this logic for multiple teams, or for multiple
schedules for one team, let's instead have the caller provide this as
a field in the `DesignMeetings` struct.