Add triage agenda generator for T-style
T-style has decided to run triage meetings in a similar way to T-lang.
Let's create a triage agenda generator for T-style that is similar to
the one in use for T-lang.

There are two main differences between the agenda for T-style and
T-lang.  One is that T-style does a "check-in" and "check-out" during
each meeting, so we have added that near the top and at the bottom of
the T-style triage agenda template.

The other difference is that T-style doesn't have a separate planning
meeting each month.  T-lang uses this planning meeting to schedule the
design meetings for the upcoming month.  Consequently, the T-lang
triage meeting agendas only show the scheduled meetings, and the
T-lang planning meeting agendas only show the proposed meetings.  For
T-style, we expect that this planning will occur during the triage
meetings, so we include both the proposed and the scheduled design
meetings on each T-style triage agenda.