Fix `proposed_fcp` parsing
When rendering queries, there's special-case handling for queries
named `proposed_fcp`.  These queries cause us to make a request to
`https://rfcbot.rs/api/all` and to parse that result.  We were getting
an opaque failure here:

```
Error: error decoding response body: invalid type: integer `...`, expected a string at line 1 column ...

Caused by:
    invalid type: integer `...`, expected a string at line 1 column ...
```

As it turns out, the reason for that failure is that the
`fk_milestone` field is of integer rather than string type.

Looking at the `rfcbot` repo, it doesn't seem that this has changed
recently, so we're uncertain why this failure occurred all of a
sudden.